Rename: 게임 로직과 관련된 객체를 baseball.baseball로 이동
