+++ src/main/java/baseball/baseball/Digits.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.baseball;
 
 import java.util.List;
